Update map plugin (#7)
* map: Replace deprecated stock icons with named symbolic icons

* map: Use a non-deprecated image icon

"gnome-mime-image" requires the deprecated gnome-icon-theme
package. Use "image-x-generic" instead, which is part of adwaita-icon-theme.

https://bugzilla.gnome.org/show_bug.cgi?id=753761

* map: Use location marker icon

Instead of showing the generic "image-x-generic" icon for a marker
on a map, use "mark-location" as the icon. This is what other GNOME
apps do, i.e. gnome-maps.

https://bugzilla.gnome.org/show_bug.cgi?id=771945

* map: Fall back to old marker icon if the new one isn't available

mark-location doesn't appear to be a "standard" icon yet,
so try falling back to image-x-generic as before if it is not available.

https://bugzilla.gnome.org/show_bug.cgi?id=771945

* map: Slowly move to selected image instead of a quick jump

When navigating through images while showing the map, the center of the map
now changes slowly from the current position to the newly selected image.

https://bugzilla.gnome.org/show_bug.cgi?id=771948

* map: Use meaningful initial zoom level

Just be sure that all given markers are visible on the initial zoom level
but on the other hand, don't show the whole world.

https://bugzilla.gnome.org/show_bug.cgi?id=771976

* map: Only set initial zoom if markers are actually present

Otherwise it would just zoom somewhere in the Atlantic.

https://bugzilla.gnome.org/show_bug.cgi?id=771976

* map: Set move duration to 1sec

Default moving duration "500 * zoom_level / 2.0" is too slow for the map
plugin purpose. The center of map now move to the next image position in
1sec regardless of zoom_level, which is similar animation duration to
other typical map apps/web apps.

Co-authored-by: Felix Riemann <friemann@gnome.org>
Co-authored-by: Thomas Bechtold <tbechtold@suse.com>